test(invariants-v2 PR-fix): move invariants-config tests to src/config
CodeRabbit PR #1459 finding 3 (MEDIUM): per the repo's co-located-tests
convention (`foo.test.ts` alongside `foo.ts`), tests targeting
`exarchos-config-schema.ts` belong in `src/config/`, not
`src/architecture/`.

The canonical location `src/config/exarchos-config-schema.test.ts`
already exists with broader schema coverage (test/typecheck/install
fields, handoffLint, cli), so the 6 Wave B1 invariants describe-block
tests are merged into that file (single test file per source module)
rather than overwriting it.

- Merge `Wave B1 — invariants.devCatalog` describe block into the
  canonical test file (preserves all 6 assertions verbatim).
- Update import path from `../config/exarchos-config-schema.js` to
  `./exarchos-config-schema.js`.
- Delete the misplaced source file.

All 27 tests pass at the new location (21 pre-existing + 6 merged).
